+++ src/services/apimap-field-builder.js
@@ -149,7 +149,7 @@ function ApimapFieldBuilder(model, column, options) {
       schema.enums = column.values;
     }
 
-    // IT WORKS.
+    // NOTICE: Create enums from sub-type (for ['Enum'] type).
     if ((Array.isArray(schema.type) && schema.type[0] === 'Enum')) {
       schema.enums = column.type.type.values;
     }