+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## RELEASE 6.1.2 - 2020-05-06
 ### Changed
 - Technical - Patch CI configuration warnings.